Fix driver valid signal generation
The drivers throughout this project work in a non-spec way. This commit fixes only one occasion - I will fix more as I encounter them. Essentially, when there is a valid-ready handshake, the drivers hold valid high for one clock cycle after it. According to spec, valid cannot be disasserted if ready is low.