Revert "Decouple CD reception to allow parallel ignore"
This reverts commit 19305e71bb560368fe362280e264463aaf2ca0e9.

The scenario occurs only when a master transmits a clean data transfer for a CleanInvalid snoop transaction. I decided to revert this commit, because this scenario is caused by master behavior that is explicitly not recommended by the specification. The spec recommends responding with DataTransfer only for Dirty data, which will be written back to memory by the interconnect.