Fix accumulating outstanding txns in coherency check
These were caused by transactions which modified cache lines by snooping but did not write their own cache. This did not cause an entry in the logs with initiator=1. Solution was to create an entry which only specifies timestamp, address, and initiator=1 to indicate the retirement of transactions that do not modify the cache.